+++ src/attribute/utils/hasNoSetter.ts
@@ -1,5 +1,8 @@
 import { noSetter } from "../const.js";
 
+/**
+ * @deprecated
+ */
 const fn = (attr: string): boolean => {
   return noSetter.includes(attr);
 };