+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Fixed
+
+- add try-catch block in attribute setter
+
 ## 1.0.8 - 2023-03-31
+
 ### Added
+
 - add `hasNoSetter()`.
 - started changelog